DE/data-pipelines/functions/training-data-ingestion/requirements.txt
# Azure Functions
azure-functions>=1.18.0

# Azure Storage (Data Lake Gen2 & Blob Storage)
azure-storage-file-datalake>=12.19.0
azure-storage-blob>=12.19.0
azure-identity>=1.15.0

# Azure Key Vault (for secrets management)
azure-keyvault-secrets>=4.8.0

# Service Bus
azure-servicebus>=7.11.0

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0

# Utilities
python-dateutil>=2.8.2
python-dotenv>=1.0.0
